ok, try for a directory guard

+++ index.js
@@ -30,16 +30,19 @@ async function uploadDocs() {
   const globber = await glob.create(patterns.join('\n'));
   var promises = [];
   for await (const file of globber.globGenerator()) {
-    // no option, so this is a buffer.
-    // if aws no likey, add , 'utf8' to the param args
-    let fileContent = fs.readFileSync(file);
-    // // https://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/S3.html#putObject-property
-    let promise =  s3.putObject({
-      Body: fileContent,
-      Bucket: bucket,
-      Key: `${projectName}/${file}`
-    });
-    promises.push(promise);
+    // if this file exists and is a file
+    if(fs.existsSync(file) && fs.lstatSync(dirPath).isFile()) {
+      // no option, so this is a buffer.
+      // if aws no likey, add , 'utf8' to the param args
+      let fileContent = fs.readFileSync(file);
+      // // https://docs.aws.amazon.com/AWSJavaScriptSDK/latest/AWS/S3.html#putObject-property
+      let promise =  s3.putObject({
+        Body: fileContent,
+        Bucket: bucket,
+        Key: `${projectName}/${file}`
+      });
+      promises.push(promise);
+    }
   }
 
   await Promise.all(promises);
